pyproject.toml
[tool.poetry]
name = "local-talking-llm"
version = "0.1.0"
description = ""
authors = ["Duy Huynh <vndee.huynh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
requests = "^2.31.0"
rich = "^13.7.1"
sqlalchemy = "^2.0.29"
torch = "^2.2.1"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
sounddevice = "^0.4.6"
suno-bark = {git = "https://github.com/suno-ai/bark.git"}
speechrecognition = "^3.10.1"
pyaudio = "^0.2.14"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
